requirements.txt
customtkinter
Pillow
requests
charset-normalizer
